Made Converters null-safe Added tests for businesspartners

+++ src/main/java/net/tomczek/invoice/manager/server/services/filestorage/LocalFileStorageService.java
@@ -13,6 +13,8 @@ import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 /**
@@ -45,7 +47,8 @@ public class LocalFileStorageService implements IFileStorageService {
         fileDAO.setFilename(fileName);
         LocalFileStorageFile savedFile = localFileStorageFileRepository.save(fileDAO);
 
-        String path = properties.getStoragePath() + "/" + savedFile.getId() + ".pdf";
+        String extension = fileName.substring(fileName.lastIndexOf("."));
+        String path = properties.getStoragePath() + "/" + savedFile.getId() + extension;
         logger.debug("Storing file with id [{}] in: [{}]", savedFile.getId(), path);
 
         File file = new File(path);
@@ -65,7 +68,8 @@ public class LocalFileStorageService implements IFileStorageService {
         Optional<LocalFileStorageFile> fileOptional = localFileStorageFileRepository.findById(fileId);
         if (fileOptional.isPresent()) {
             LocalFileStorageFile fileNameMapping = fileOptional.get();
-            String path = properties.getStoragePath() + "/" + fileId + ".pdf";
+            String extension = fileNameMapping.getFilename().substring(fileNameMapping.getFilename().lastIndexOf("."));
+            String path = properties.getStoragePath() + "/" + fileId + extension;
             logger.debug("Retrieving file with id [{}] from: [{}]", fileId, path);
             return new FileWithContent(fileNameMapping.getId(), fileNameMapping.getFilename(), Files.readAllBytes(Paths.get(path)));
         }
@@ -74,16 +78,31 @@ public class LocalFileStorageService implements IFileStorageService {
 
     @Override
     public boolean deleteFile(int fileId) throws Exception {
-        localFileStorageFileRepository.findById(fileId).ifPresent(localFileStorageFileRepository::delete);
-        String path = properties.getStoragePath() + "/" + fileId + ".pdf";
+        Optional<LocalFileStorageFile> storedFile = localFileStorageFileRepository.findById(fileId);
+        if (storedFile.isEmpty()) {
+            return false;
+        }
+        localFileStorageFileRepository.delete(storedFile.get());
+        String extension = storedFile.get().getFilename().substring(storedFile.get().getFilename().lastIndexOf("."));
+
+        String path = properties.getStoragePath() + "/" + fileId + extension;
         logger.debug("Deleting file with id [{}] from: [{}]", fileId, path);
         return new File(path).delete();
     }
 
     @Override
     public boolean fileExists(int fileId) throws Exception {
-        String path = properties.getStoragePath() + "/" + fileId + ".pdf";
-        return new File(path).exists();
+        File storageDirectory = new File(properties.getStoragePath());
+        if (!storageDirectory.exists()) {
+            return false;
+        }
+
+        File[] files = storageDirectory.listFiles((dir, name) -> name.startsWith(fileId + "."));
+        if (Objects.isNull(files) || files.length == 0) {
+            return false;
+        }
+
+        return files[0].exists();
     }
 
     private boolean initStorageDirectoryIfNecessary() {
